Add UI ListChooser
A ListChooser has a collection of ListElements, each of which is a
String drawn to the String.  The user scrolls through each ListElement
using the arrow keys, and the current ListElement is drawn to be a
different color.  When enter is pressed, the currently selected
ListElement's trigger method is called.  ListElement, therefore, is
an abstract class so that the trigger method is implemented on
construction.